Change in migration ticket and movie

+++ db/migrate/20220110090014_create_tickets.rb
@@ -3,8 +3,8 @@ class CreateTickets < ActiveRecord::Migration[7.0]
     create_table :tickets do |t|
       t.string :seat
       t.string :type
-      t.float :price
-
+      t.decimal :price
+      
       t.timestamps
     end
   end
