+++ db/migrate/20220110090402_create_movies.rb
@@ -3,7 +3,7 @@ class CreateMovies < ActiveRecord::Migration[7.0]
     create_table :movies do |t|
       t.string :description
       t.string :title
-      t.float :duration
+      t.integer :duration
 
       t.timestamps
     end
